requirements.txt
pillow==11.0.0
pynput==1.7.7
six==1.16.0
